ProseMirror.Net.UnitTests/Nodes/BulletListTests.cs
using FluentAssertions;
using HtmlAgilityPack;
using ProseMirror.Net.Nodes;
using System;
using Xunit;

namespace ProseMirror.Net.UnitTests.Nodes
{
    public class BulletListTests
    {
        [Fact]
        public void Matches_ShouldReturnTrue_WhenHtmlTagUl()
        {
            // arrange
            var doc = new HtmlDocument();
            doc.LoadHtml("<ul><li>This should be bullet list</li></ul>");

            // act
            var matches = new BulletList().Matches(doc.DocumentNode.ChildNodes[0]);

            // assert
            matches.Should().BeTrue();
        }

        [Fact]
        public void Matches_ShouldReturnFalse_WhenHtmlTagNotUl()
        {
            // arrange
            var doc = new HtmlDocument();
            doc.LoadHtml("<span>This should NOT be bullet list</span>");

            // act
            var matches = new BulletList().Matches(doc.DocumentNode.ChildNodes[0]);

            // assert
            matches.Should().BeFalse();
        }

        [Fact]
        public void Data_ShouldReturnBulletListMark_WhenValidNode()
        {
            // arrange
            var doc = new HtmlDocument();
            doc.LoadHtml("<ul><li>This should be bullet list</li></ul>");

            // act
            var data = new BulletList().Data(doc.DocumentNode.ChildNodes[0]);

            // assert
            data.TypeEnum.Should().Be(Model.NodeType.BulletList);
        }

        [Fact]
        public void Data_ShouldThrow_WhenInvalidNode()
        {
            // arrange
            var doc = new HtmlDocument();
            doc.LoadHtml("<span>This should be bullet list</span>");

            // act
            var act = () => new BulletList().Data(doc.DocumentNode.ChildNodes[0]);

            // assert
            act.Should().Throw<InvalidOperationException>();
        }
    }
}
